backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
sqlmodel==0.0.14
google-api-python-client==2.108.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
httpx==0.25.1
python-multipart==0.0.6
aiosqlite==0.19.0
